Add files via upload
This Pull Request heavily modifies the Petrochemical and Hydrocarbon section of the MV Quests, in order to give a better sense of progression and provide additional material. As a player, this section was deeply confusing and unwieldy to navigate when I first entered MV. I hope that this new quest line will make it easier for players to understand the intricacies of the distillation and petrochemical mechanics. New quests have been added to more easily split the Natural and Oil-based routes, and a quest for Wood Tar has been added. Desulfurization is now its own quest, and this should hopefully prevent the repetitive explanations of desulfurization, while offering more details.